# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A lightweight, containerized REST API service that provides video and audio download and streaming functionalities using yt-dlp and ffmpeg.",
        "title": "GoStreamPuller API",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/download/audio": {
            "post": {
                "description": "Downloads an audio file from a given URL to the server's download directory.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "download"
                ],
                "summary": "Download an audio file",
                "parameters": [
                    {
                        "description": "Audio download request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DownloadAudioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audio downloaded successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.DownloadAudioResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload or missing URL",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error during audio download",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/download/audio/{filename}": {
            "get": {
                "description": "Serves an audio file from the server's download directory given its filename.",
                "produces": [
                    "audio/mpeg"
                ],
                "tags": [
                    "download"
                ],
                "summary": "Serve a downloaded audio file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filename of the audio to serve",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully served audio file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Missing filename",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/stream/audio": {
            "post": {
                "description": "Streams an audio file directly from the source URL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "audio/mpeg"
                ],
                "tags": [
                    "stream"
                ],
                "summary": "Stream an audio file",
                "parameters": [
                    {
                        "description": "Audio stream request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.StreamAudioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully streamed audio",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload or missing URL",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error during audio streaming",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.DownloadAudioRequest": {
            "type": "object",
            "properties": {
                "bitrate": {
                    "type": "string"
                },
                "codec": {
                    "type": "string"
                },
                "outputFormat": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "handler.DownloadAudioResponse": {
            "type": "object",
            "properties": {
                "filePath": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "videoInfo": {
                    "description": "Re-use VideoInfo for audio metadata",
                    "allOf": [
                        {
                            "$ref": "#/definitions/service.VideoInfo"
                        }
                    ]
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.StreamAudioRequest": {
            "type": "object",
            "properties": {
                "bitrate": {
                    "type": "string"
                },
                "codec": {
                    "type": "string"
                },
                "outputFormat": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "service.VideoInfo": {
            "type": "object",
            "properties": {
                "duration": {
                    "description": "in seconds",
                    "type": "integer"
                },
                "ext": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "original_url": {
                    "type": "string"
                },
                "thumbnail": {
                    "description": "URL to thumbnail",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "upload_date": {
                    "description": "YYYYMMDD",
                    "type": "string"
                },
                "uploader": {
                    "type": "string"
                }
            }
        }
    }
}